Provide abstract interface to p::d::Triangulation::load()/save()
