+++ doc/news/changes/minor/20210118PasqualeAfrica
@@ -1,0 +1,3 @@
+Improved: provide abstract interface to p::d::Triangulation::load()/save().
+<br>
+(Pasquale Claudio Africa, 2021/01/18)
